+++ app/mailers/test_result_mailer.rb
@@ -8,7 +8,9 @@ class TestResultMailer < ApplicationMailer
   # @return [nil]
   def spec_failed_email(history)
     @history = history
-    mail(to: Rails.application.credentials.admin_email, subject: t(:mailer_something_wrong_with_test_title))
+    Rails.application.credentials.admin_emails.each do |email|
+      mail(to: email, subject: t(:mailer_something_wrong_with_test_title))
+    end
   end
 
   # This email will be send if there is not a single `it`
@@ -17,6 +19,8 @@ class TestResultMailer < ApplicationMailer
   # @return [nil]
   def spec_no_tests_executed_email(history)
     @history = history
-    mail(to: Rails.application.credentials.admin_email, subject: t(:mailer_not_a_single_test_title))
+    Rails.application.credentials.admin_emails.each do |email|
+      mail(to: email, subject: t(:mailer_not_a_single_test_title))
+    end
   end
 end