+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changes
 
 * Remove several unused constants
+* `Rails.application.credentials.admin_email` is now an array of
+  `Rails.application.credentials.admin_emails` to have several admins
 
 ## v1.30.0 (2022-10-31)